requirements.txt
google-adk>=0.2.0
fastapi>=0.115.0
uvicorn>=0.23.0
pydantic>=2.0.0
zoneinfo>=0.2.1; python_version < '3.9'
python-multipart>=0.0.6
flask>=2.0.0
requests>=2.31.0
google-cloud-aiplatform[adk,agent_engines]==1.92.0
python-dotenv>=1.0.0